feat: move browser management logic from browser service to new browser class
