+++ src/BlazorReports/Services/BrowserServices/Problems/BrowserProblem.cs
@@ -1,0 +1,6 @@
+namespace BlazorReports.Services.BrowserServices.Problems;
+
+/// <summary>
+/// Represents a problem that occurs when the browser fails to start
+/// </summary>
+public readonly record struct BrowserProblem;
