+++ src/BlazorReports/Services/BrowserServices/Problems/OperationCancelledProblem.cs
@@ -3,4 +3,4 @@
 /// <summary>
 /// Represents a problem when the operation is cancelled
 /// </summary>
-public readonly record struct OperationCancelledProblem();
+public readonly record struct OperationCancelledProblem;
