+++ src/BlazorReports/Services/BrowserServices/Problems/ServerBusyProblem.cs
@@ -3,4 +3,4 @@ namespace BlazorReports.Services.BrowserServices.Problems;
 /// <summary>
 /// Represents a problem that occurs when the server is busy.
 /// </summary>
-public readonly record struct ServerBusyProblem();
+public readonly record struct ServerBusyProblem;
