+++ src/BlazorReports/Services/ReportService.cs
@@ -48,7 +48,7 @@ public sealed class ReportService : IReportService, IAsyncDisposable
   /// <typeparam name="T"> The component to use in the report </typeparam>
   /// <typeparam name="TD"> The type of data to use in the report </typeparam>
   /// <returns> The generated report </returns>
-  public async ValueTask<OneOf<Success, ServerBusyProblem, OperationCancelledProblem>> GenerateReport<T, TD>(
+  public async ValueTask<OneOf<Success, ServerBusyProblem, OperationCancelledProblem, BrowserProblem>> GenerateReport<T, TD>(
     PipeWriter pipeWriter, TD data, CancellationToken cancellationToken = default)
     where T : ComponentBase where TD : class
   {
@@ -76,7 +76,15 @@ public sealed class ReportService : IReportService, IAsyncDisposable
       return output.ToHtmlString();
     });
 
-    return await _browserService.PrintReportFromBrowser(pipeWriter, html, _reportRegistry.DefaultPageSettings,
+    var result = await _browserService.GetBrowser();
+    var hasBrowserStartProblem =
+      result.TryPickT1(out var browserStartProblem, out var browser);
+    if (hasBrowserStartProblem)
+    {
+      return browserStartProblem;
+    }
+
+    return await browser.GenerateReport(pipeWriter, html, _reportRegistry.DefaultPageSettings,
       cancellationToken);
   }
 
@@ -89,7 +97,7 @@ public sealed class ReportService : IReportService, IAsyncDisposable
   /// <param name="cancellationToken"> The cancellation token </param>
   /// <typeparam name="T"> The type of data to use in the report </typeparam>
   /// <returns> The generated report </returns>
-  public async ValueTask<OneOf<Success, ServerBusyProblem, OperationCancelledProblem>> GenerateReport<T>(
+  public async ValueTask<OneOf<Success, ServerBusyProblem, OperationCancelledProblem, BrowserProblem>> GenerateReport<T>(
     PipeWriter pipeWriter, BlazorReport blazorReport, T? data,
     CancellationToken cancellationToken = default) where T : class
   {
@@ -141,7 +149,16 @@ public sealed class ReportService : IReportService, IAsyncDisposable
 
     var pageSettings = blazorReport.PageSettings ?? _reportRegistry.DefaultPageSettings;
 
-    return await _browserService.PrintReportFromBrowser(pipeWriter, html, pageSettings, cancellationToken);
+    var result = await _browserService.GetBrowser();
+    var hasBrowserStartProblem =
+      result.TryPickT1(out var browserStartProblem, out var browser);
+    if (hasBrowserStartProblem)
+    {
+      return browserStartProblem;
+    }
+
+    return await browser.GenerateReport(pipeWriter, html, pageSettings, cancellationToken);
+
   }
 
   /// <summary>
@@ -151,7 +168,7 @@ public sealed class ReportService : IReportService, IAsyncDisposable
   /// <param name="blazorReport"> The report to generate </param>
   /// <param name="cancellationToken"> The cancellation token </param>
   /// <returns> The generated report </returns>
-  public ValueTask<OneOf<Success, ServerBusyProblem, OperationCancelledProblem>> GenerateReport(PipeWriter pipeWriter,
+  public ValueTask<OneOf<Success, ServerBusyProblem, OperationCancelledProblem, BrowserProblem>> GenerateReport(PipeWriter pipeWriter,
     BlazorReport blazorReport, CancellationToken cancellationToken = default)
   {
     return GenerateReport<object>(pipeWriter, blazorReport, null, cancellationToken);
